+++ Proline-Databases/src/main/resources/dbscripts/msi/postgresql/V0_1__init_msidb.sql
@@ -768,7 +768,7 @@ CREATE TABLE public.peptide_instance (
                 protein_match_count INTEGER NOT NULL,
                 protein_set_count INTEGER NOT NULL,
                 selection_level INTEGER NOT NULL,
-                elution_time DOUBLE PRECISION,
+                elution_time REAL,
                 serialized_properties TEXT,
                 best_peptide_match_id INTEGER NOT NULL,
                 peptide_id INTEGER NOT NULL,
